Refactor DCIM forms
