+++ netbox/dcim/forms/__init__.py
@@ -1,0 +1,10 @@
+from .fields import *
+from .models import *
+from .filtersets import *
+from .object_create import *
+from .object_import import *
+from .bulk_create import *
+from .bulk_edit import *
+from .bulk_import import *
+from .connections import *
+from .formsets import *
